Minor Tweaks
Changes to the crossover method to return two children instead of one child - allows for greater exploration.

Tournament selection is now available as an option for the selection criteria.

Fitness function now includes wasted space as a criteria. Circles will be penalized for being away from the center of the container.

Code to visualise change in fitness over generations