UI improvements and test cases
Final improvements to the UI with some nice features such as "run all for test case". Console no longer prints to the GUI as this was slowing down execution.

Added initial test cases to dictionary. The test cases are automatically pulled into the UI from the dict with no additional work so extensible